requirements.txt
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.18.0
numpy>=1.21.0
asyncio
typing
pathlib
dataclasses 